Move helpers to lib folder
