+++ payper-app/src/integrations/subgraph/hooks/use-get-articles-from-journalist.ts
@@ -1,46 +1,79 @@
-import { ArticleData } from '@/types';
-import { ApolloClient, gql } from "@apollo/client";
+import {  ApolloClient, gql } from "@apollo/client";
+import { useEffect, useState } from 'react';
+import {ArticleData} from '@/types';
 
 interface UseGetArticlesFromJournalistParams {
   client: ApolloClient<object>;
-  journalistId: number;
+  journalistId: string;
+}
+
+interface UseGetArticlesFromJournalistResponse {
+  isLoading: Boolean;
+  articles: ArticleData[];
 }
 
 /**
 * Fetches the information about the articles of a journalist
 */
-const useGetArticleFromJournalist = async ({
+const useGetArticlesFromJournalist = ({
   client,
   journalistId,
-}: UseGetArticlesFromJournalistParams): Promise<ArticleData[]> => {
-  const result = await client.query({
-    query: gql`
+}: UseGetArticlesFromJournalistParams): UseGetArticlesFromJournalistResponse => {
+  const [articles, setArticles] = useState<ArticleData[]>([]);
+  const [isLoading, setIsLoading] = useState<Boolean>(true);
+
+useEffect(() => {
+  async function fetchData() {
+    try {
+      const result = await client.query({
+        query: gql`
           query {
             articles(where: {journalist: "${journalistId}"}) {
               id
               name
               freeContent
-              journalist
+              encryptedUrl
+              totalRating
+              amountOfRatings
+              price
+              totalPaymentReceived
+              date
+              newsType
             }
           }
         `,
-  });
-  const data = result.data.articles;
-  const articles: ArticleData[] = data ? data.map((articleData: any) => ({
-    id: articleData.id,
-    journalist: articleData.journalist,
-    name: articleData.name,
-    freeContent: articleData.freeContent,
-    encryptedUrl: articleData.encryptedUrl,
-    totalRating: articleData.totalRating,
-    amountOfRatings: articleData.amountOfRatings,
-    price: articleData.price,
-    totalPaymentReceived: articleData.totalPaymentReceived,
-    date: articleData.date,
-    newsType: articleData.newsType,
-  })) : [];
+      });
+      const data = result.data.articles;
 
-  return articles;
+      const responseArticles: ArticleData[] = data 
+        ? data.map((articleData: any) => ({
+          id: articleData.id,
+          name: articleData.name,
+          freeContent: articleData.freeContent,
+          encryptedUrl: articleData.encryptedUrl,
+          totalRating: articleData.totalRating,
+          amountOfRatings: articleData.amountOfRatings,
+          price: articleData.price,
+          totalPaymentReceived: articleData.totalPaymentReceived,
+          date: articleData.date,
+          newsType: articleData.newsType,
+        })) 
+        : [];
+      setArticles(responseArticles);
+      setIsLoading(false);
+    } catch (error) {
+      console.error('Error fetching data:', error);
+      setIsLoading(false);
+    }
+  }
+
+  fetchData();
+}, []);
+    
+  return {
+    isLoading,
+    articles,
+  };
 };
 
-export default useGetArticleFromJournalist;
+export default useGetArticlesFromJournalist;
