+++ payper-app/src/lib/calculate-dashboard-information.ts
