+++ payper-app/src/lib/calculate-rating.ts
